move batch handlers to predictive
