+++ chirho/robust/internals/utils.py
@@ -1,16 +1,14 @@
 import contextlib
 import functools
 import math
-from typing import Any, Callable, Dict, Generic, Mapping, Optional, Tuple, TypeVar
+from typing import Any, Callable, Dict, Mapping, Optional, Tuple, TypeVar
 
 import pyro
 import torch
 from typing_extensions import Concatenate, ParamSpec
 
 from chirho.indexed.handlers import add_indices
-from chirho.indexed.ops import IndexSet, get_index_plates, indices_of
-from chirho.observational.handlers.condition import Observations
-from chirho.robust.ops import Point
+from chirho.indexed.ops import IndexSet, get_index_plates
 
 P = ParamSpec("P")
 Q = ParamSpec("Q")
@@ -213,79 +211,3 @@ def site_is_delta(msg: dict) -> bool:
     while hasattr(d, "base_dist"):
         d = d.base_dist
     return isinstance(d, pyro.distributions.Delta)
-
-
-class BatchedLatents(pyro.poutine.messenger.Messenger):
-    """
-    Effect handler that adds a fresh batch dimension to all latent ``sample`` sites.
-    Similar to wrapping a Pyro model in a ``pyro.plate`` context, but uses the machinery
-    in ``chirho.indexed`` to automatically allocate and track the fresh batch dimension
-    based on the ``name`` argument to ``BatchedLatents`` .
-
-    .. warning:: Must be used in conjunction with :class:`~chirho.indexed.handlers.IndexPlatesMessenger` .
-
-    :param int num_particles: Number of particles to use for parallelization.
-    :param str name: Name of the fresh batch dimension.
-    """
-
-    num_particles: int
-    name: str
-
-    def __init__(self, num_particles: int, *, name: str = "__particles_mc"):
-        assert num_particles > 0
-        assert len(name) > 0
-        self.num_particles = num_particles
-        self.name = name
-        super().__init__()
-
-    def _pyro_sample(self, msg: dict) -> None:
-        if (
-            self.num_particles > 1
-            and msg["value"] is None
-            and not pyro.poutine.util.site_is_factor(msg)
-            and not pyro.poutine.util.site_is_subsample(msg)
-            and not site_is_delta(msg)
-            and self.name not in indices_of(msg["fn"])
-        ):
-            msg["fn"] = unbind_leftmost_dim(
-                msg["fn"].expand((1,) + msg["fn"].batch_shape),
-                self.name,
-                size=self.num_particles,
-            )
-
-
-class BatchedObservations(Generic[T], Observations[T]):
-    """
-    Effect handler that takes a dictionary of observation values for ``sample`` sites
-    that are assumed to be batched along their leftmost dimension, adds a fresh named
-    dimension using the machinery in ``chirho.indexed``, and reshapes the observation
-    values so that the new ``chirho.observational.observe`` sites are batched along
-    the fresh named dimension.
-
-    Useful in combination with ``pyro.infer.Predictive`` which returns a dictionary
-    of values whose leftmost dimension is a batch dimension over independent samples.
-
-    .. warning:: Must be used in conjunction with :class:`~chirho.indexed.handlers.IndexPlatesMessenger` .
-
-    :param Point[T] data: Dictionary of observation values.
-    :param str name: Name of the fresh batch dimension.
-    """
-
-    name: str
-
-    def __init__(self, data: Point[T], *, name: str = "__particles_data"):
-        assert len(name) > 0
-        self.name = name
-        super().__init__(data)
-
-    def _pyro_observe(self, msg: dict) -> None:
-        super()._pyro_observe(msg)
-        if msg["kwargs"]["name"] in self.data:
-            rv, obs = msg["args"]
-            event_dim = (
-                len(rv.event_shape)
-                if hasattr(rv, "event_shape")
-                else msg["kwargs"].get("event_dim", 0)
-            )
-            batch_obs = unbind_leftmost_dim(obs, self.name, event_dim=event_dim)
-            msg["args"] = (rv, batch_obs)
